Add files via upload
ODBC file was used to connect the Microsoft Access database to Microsoft SQL Server. Opening the access file you should see that each table has been recognized as a database object (DBO) and that is the table which is linked to SQL server. This allows to the database to be stored on the cloud but still allows for control and manipulation through Microsoft Access.